// src/entities/Profile/model/slice/profileSlice.ts
import { PayloadAction, createSlice } from '@reduxjs/toolkit'
import { Profile, type ProfileSchema } from '../types/profile'
import { fetchProfileData } from '../services/fetchProfileData/fetchProfileData'
import { Country, Currency } from 'shared/const/common'

const initialState: ProfileSchema = {
  data: {
    first: '',
    lastname: '',
    age: 0,
    currency: Currency.USD,
    country: Country.Ukraine,
    city: '',
    username: '',
    avatar: '',
  },
  isLoading: false,
  error: '',
  readonly: true,
}

export const profileSlice = createSlice({
  name: 'profile',
  initialState,
  reducers: {
  },
  extraReducers: (builder) => {
    builder
      .addCase(fetchProfileData.pending, (state) => {
        state.error = undefined
        state.isLoading = true
      })
      .addCase(fetchProfileData.fulfilled, (state, action: PayloadAction<Profile>) => {
        state.isLoading = false
        state.data = action.payload
      })
      .addCase(fetchProfileData.rejected, (state, action) => {
        state.isLoading = false
        state.error = action.payload
      })
  }
})

// Action creators are generated for each case reducer function
export const { actions: profileActions, reducer: profileReducer } = profileSlice
